% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents.R
\name{agent}
\alias{agent}
\title{View an Agent}
\usage{
agent(client, agent_id, agents_path = "/api/v2/agents",
  ticket_scope_lookup = agent_ticket_scope)
}
\arguments{
\item{client}{The Freshdesk API client object (see \code{\link{freshdesk_client}}).}

\item{agent_id}{The ID of the agent you would like to view.}

\item{agents_path}{The path of the agents API. Defaults to \code{/api/v2/agents}.
You should not need to change the default value.}

\item{ticket_scope_lookup}{Optional dataframe of ticket scope and associated values.
Defaults to \code{\link{agent_ticket_scope}} which is defined in the package.}
}
\value{
A list with the attributes of the agent.
}
\description{
\code{agent} queries the Freshdesk API and returns data related to an agent.
}
\details{
This function queries the Freshdesk API based on agent id to view the details
regarding this agent.
}
\examples{
\dontrun{
fc <- freshdesk_client("your-domain", "your-api-key")

# view agent with id = 42001318958
a <- agent(fc, 42001318958)
}
}
