% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agent_ticket_scope}
\alias{agent_ticket_scope}
\title{Lookup table for ticket scope}
\format{A data frame with 3 rows and 2 variables:
\describe{
  \item{Scope}{the ticket scope}
  \item{Value}{the numeric value returned by the API}
}}
\source{
\url{https://developers.freshdesk.com/api/}
}
\usage{
agent_ticket_scope
}
\description{
A dataset containing the ticket scope and value for valid
ticket scope
}
\keyword{datasets}
