% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents.R
\name{agents}
\alias{agents}
\title{View a list of Agents}
\usage{
agents(client, agents_path = "/api/v2/agents", max_records = Inf,
  ticket_scope_lookup = agent_ticket_scope, date_fields = agent_date_fields)
}
\arguments{
\item{client}{The Freshdesk API client object (see \code{\link{freshdesk_client}}).}

\item{agents_path}{The path of the agents API. Defaults to \code{/api/v2/agents}.
You should not need to change the default value.}

\item{max_records}{Specifies the maximum number of records to return.}

\item{ticket_scope_lookup}{Optional dataframe of ticket scope and associated values.
Defaults to \code{\link{agent_ticket_scope}} which is defined in the package.}

\item{date_fields}{Fields returned by the Freshdesk API that are date fields. Defaults
to \code{\link{agent_date_fields}} which is defined in the package.}
}
\value{
A data frame of agents or \code{NULL} if there are no agents.
}
\description{
\code{tickets} queries the Freshdesk API and returns a data frame of agent data.
}
\details{
This function queries the Freshdesk API to view the details regarding multiple
agents
}
\examples{
\dontrun{
fc <- freshdesk_client("your-domain", "your-api-key")

# view agents
a <- agents(fc)
}
}
