% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies.R
\name{companies}
\alias{companies}
\title{View a list of Companies}
\usage{
companies(client, companies_path = "/api/v2/companies", max_records = Inf,
  date_fields = company_date_fields)
}
\arguments{
\item{client}{The Freshdesk API client object (see \code{\link{freshdesk_client}}).}

\item{companies_path}{The path of the companies API. Defaults to \code{/api/v2/companies}.
You should not need to change the default value.}

\item{max_records}{Specifies the maximum number of records to return.}

\item{date_fields}{Fields returned by the Freshdesk API that are date fields. Defaults
to \code{\link{company_date_fields}} which is defined in the package.}
}
\value{
A data frame of companies or \code{NULL} if there are no companies.
}
\description{
\code{companies} queries the Freshdesk API and returns a data frame of company data.
}
\details{
This function queries the Freshdesk API to view the details regarding multiple
companies
}
\examples{
\dontrun{
fc <- freshdesk_client("your-domain", "your-api-key")

# view companies
cos <- companies(fc)
}
}
