% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies.R
\name{company}
\alias{company}
\title{View an Company}
\usage{
company(client, company_id, companies_path = "/api/v2/companies")
}
\arguments{
\item{client}{The Freshdesk API client object (see \code{\link{freshdesk_client}}).}

\item{company_id}{The ID of the company you would like to view.}

\item{companies_path}{The path of the companies API. Defaults to \code{/api/v2/companies}.
You should not need to change the default value.}
}
\value{
A list with the attributes of the company
}
\description{
\code{agent} queries the Freshdesk API and returns data related to an comapny.
}
\details{
This function queries the Freshdesk API based on company id to view the details
regarding this company.
}
\examples{
\dontrun{
fc <- freshdesk_client("your-domain", "your-api-key")

# view compnay with id = 42001318958
co <- company(fc, 42001318958)
}
}
