% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{company_date_fields}
\alias{company_date_fields}
\title{Company date fields returned by the Freshdesk API.}
\format{A vector}
\source{
\url{https://developers.freshdesk.com/api/}
}
\usage{
company_date_fields
}
\description{
A list containing all the potential date fields that
may be returned with company data from the Freshdesk API.
}
\keyword{datasets}
