% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{contact}
\alias{contact}
\title{View a Contact}
\usage{
contact(client, contact_id, contacts_path = "/api/v2/contacts")
}
\arguments{
\item{client}{The Freshdesk API client object (see \code{\link{freshdesk_client}}).}

\item{contact_id}{The ID of the contact you would like to view.}

\item{contacts_path}{The path of the contacts API. Defaults to \code{/api/v2/contacts}.
You should not need to change the default value.}
}
\value{
A list with the attributes of the contact.
}
\description{
\code{contact} queries the Freshdesk API and returns data related to a contact.
}
\details{
This function queries the Freshdesk API based on contact id to view the details
regarding this contact.
}
\examples{
\dontrun{
fc <- freshdesk_client("your-domain", "your-api-key")

# view contact with id = 42001318958
ct <- contact(fc, 42001318958)
}
}
