% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{contacts}
\alias{contacts}
\title{View a list of Contacts}
\usage{
contacts(client, contacts_path = "/api/v2/contacts", max_records = Inf,
  date_fields = contact_date_fields)
}
\arguments{
\item{client}{The Freshdesk API client object (see \code{\link{freshdesk_client}}).}

\item{contacts_path}{The path of the agents API. Defaults to \code{/api/v2/contacts}.
You should not need to change the default value.}

\item{max_records}{Specifies the maximum number of records to return.}

\item{date_fields}{Fields returned by the Freshdesk API that are date fields. Defaults
to \code{\link{contact_date_fields}} which is defined in the package.}
}
\value{
A data frame of contacts or \code{NULL} if there are no contacts.
}
\description{
\code{contacts} queries the Freshdesk API and returns a data frame of contact data.
}
\details{
This function queries the Freshdesk API to view the details regarding multiple
contacts
}
\examples{
\dontrun{
fc <- freshdesk_client("your-domain", "your-api-key")

# view contacts
cs <- contacts(fc)
}
}
