% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{freshdesk_api}
\alias{freshdesk_api}
\title{Calls Freshdesk API}
\usage{
freshdesk_api(client, path, query = NULL, per_page = NULL, pages = Inf)
}
\arguments{
\item{client}{The Freshdesk API client object (see \code{\link{freshdesk_client}}).}

\item{path}{The API query path.}

\item{query}{API query string.}

\item{per_page}{The number of results per page. The default is 30. Values over 100 will
result in an error being raised as it wis not allowed by the Freshdesk API.}

\item{pages}{By default, all pages will be returned. Specify this value along with the
\code{per_page} parameter to limit the number of results returned.}
}
\value{
An S3 object contaitning the following attributes:
  \itemize{
    \item{\code{content}}: {the parsed content of the response.}
    \item{\code{path}}: {the API query path.}
    \item{\code{response}}: {the complete httr reponse objects.}
    \item{\code{rate_limit_remaining}}: {the API calls remaining in the current period.}
    \item{\code{rate_limit_total}}: {the total API calls for the current period.}
  }
}
\description{
\code{freshdesk_api} queries the Freshdesk API and returns a result.
}
\details{
This function queries the Freshdesk API based on a path and returns a \code{freshdesk_api}
object containing the http responses, the parsed content, and the API rate limit status. If the
results are paginated, all of the results will be returned up to the number of pages specified
(all pages by default). If the results span multiple pages, a list of http responses and the
rate limit information from the final call is returned in the \code{freshdesk_api} object along
with \strong{all} the data combined.
}
\examples{
\dontrun{
fc <- freshdesk_client("foo", "MyAPIKey")
apidata <- freshdesk_api(fc, "/api/v2/tickets/3")
apidata$rate_limit_remaining
}
}
