% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{freshdesk_api_call}
\alias{freshdesk_api_call}
\title{Calls Freshdesk API}
\usage{
freshdesk_api_call(client, path, query = NULL)
}
\arguments{
\item{client}{The Freshdesk API client object (see \code{\link{freshdesk_client}}).}

\item{path}{The API query path.}

\item{query}{API query string.}
}
\value{
An S3 object contaitning the following attributes:
  \itemize{
    \item{\code{content}}: {the parsed content of the response.}
    \item{\code{path}}: {the API query path.}
    \item{\code{response}}: {the complete httr reponse object.}
    \item{\code{rate_limit_remaining}}: {the API calls remaining in the current period.}
    \item{\code{rate_limit_total}}: {the total API calls for the current period.}
  }
}
\description{
\code{freshdesk_api_call} makes a single query the Freshdesk API and returns a result.
}
\details{
This function queries the Freshdesk API based on a path and returns a \code{freshdesk_api}
object containing the http response, the parsed content, and the API rate limit status.
}
\examples{
\dontrun{
fc <- freshdesk_client("foo", "MyAPIKey")
apidata <- freshdesk_api(fc, "/api/v2/tickets/3")
apidata$rate_limit_remaining
}
}
