% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{freshdesk_client}
\alias{freshdesk_client}
\title{Creates a Freshdesk API client}
\usage{
freshdesk_client(domain, api_key, password = "x", check_connection = FALSE,
  check_api_path = "/api/v2/settings/helpdesk")
}
\arguments{
\item{domain}{The Freshdesk domain, e.g. \code{https://<your-domain>.freshdesk.com}.}

\item{api_key}{Your Freshdesk api key. This can be found under \strong{Your API Key} on your
Freshdesk \strong{Profile Settings} page. Otherwise set this to your user name to use basic
authentication.}

\item{password}{Your password. Only required if you are using basic authentication.}

\item{check_connection}{If \code{TRUE}, checks internet connection and connection to the API
by querying \code{check_api_path}. \strong{Note: Checking the connection consumes API call credits}.}

\item{check_api_path}{The API path used to test the connection if \code{check_connection} is set
to \code{TRUE}.}
}
\value{
A list of your credentials. This list is required by the \code{\link{freshdesk_api}}
  function to call methods through the API.
}
\description{
\code{freshdesk_client} returns an object containing Freshdesk credentials.
}
\details{
This function takes data regarding a Freshdesk account and creates a list object
continaing these credentials. This list contains data that will be used to
authenticate on Freshdesk when using the Freshdesk API.
}
\examples{
\dontrun{
fc <- freshdesk_client("foo", "me@foo.com", "myPassword")
fc <- freshdesk_client("foo", "MyAPIKey")
}
}
