% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshdeskr.R
\docType{package}
\name{freshdeskr}
\alias{freshdeskr}
\alias{freshdeskr-package}
\title{freshdeskr: An R wrapper for the \href{https://developers.freshdesk.com/api/}{Freshdesk API}.}
\description{
The freshdeskr package provides access to Freshdesk data via the Freshdesk API in R.
}
