% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.freshdesk_api}
\alias{print.freshdesk_api}
\title{Generic print for \code{freshdesk_api} class}
\usage{
\method{print}{freshdesk_api}(x, ...)
}
\arguments{
\item{x}{freshdesk api object}

\item{...}{other arguments}
}
\description{
Prints the class name, parsed json and current rate limit information
}
