% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickets.R
\name{ticket}
\alias{ticket}
\title{View a Ticket}
\usage{
ticket(client, ticket_id, tickets_path = "/api/v2/tickets", include = NULL)
}
\arguments{
\item{client}{The Freshdesk API client object (see \code{\link{freshdesk_client}}).}

\item{ticket_id}{The ID of the ticket you would like to view.}

\item{tickets_path}{The path of the tickets API. Defaults to \code{/api/v2/tickets}.
You should not need to change the default value.}

\item{include}{Optional include query string parameters.}
}
\value{
A list with the attributes of the ticket.
}
\description{
\code{ticket} queries the Freshdesk API and returns data related to a ticket.
}
\details{
This function queries the Freshdesk API based on ticket id to view the details
regarding this ticket. You can optionally include conversation, requester, company
or extra stats data.
}
\examples{
\dontrun{
fc <- freshdesk_client("your-domain", "your-api-key")

# view ticket with id = 3 and include addtional company and requester data
t <- ticket(fc, 3, include = "requester,company")
}
}
