% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ticket_include_parameters}
\alias{ticket_include_parameters}
\title{Ticket embedding parameters by the Freshdesk API.}
\format{A vector}
\source{
\url{https://developers.freshdesk.com/api/}
}
\usage{
ticket_include_parameters
}
\description{
A list containing all the valid parametes that can be
used by \code{include} query string when extracting
tickt data from the Freshdesk API.
}
\keyword{datasets}
