% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ticket_priorities}
\alias{ticket_priorities}
\title{Lookup table for ticket priorities.}
\format{A data frame with 4 rows and 2 variables:
\describe{
  \item{Priority}{the priority name}
  \item{Value}{the numeric value returned by the API}
}}
\source{
\url{https://developers.freshdesk.com/api/}
}
\usage{
ticket_priorities
}
\description{
A dataset containing the priorty and value for valid
ticket priorities.
}
\keyword{datasets}
