% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ticket_sources}
\alias{ticket_sources}
\title{Lookup table for ticket sources.}
\format{A data frame with 7 rows and 2 variables:
\describe{
  \item{Source}{the source name}
  \item{Value}{the numeric value returned by the API}
}}
\source{
\url{https://developers.freshdesk.com/api/}
}
\usage{
ticket_sources
}
\description{
A dataset containing the source and value for valid
ticket sources.
}
\keyword{datasets}
