% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickets.R
\name{tickets}
\alias{tickets}
\title{View a list of Tickets}
\usage{
tickets(client, tickets_path = "/api/v2/tickets",
  remove_fields = c("description", "description_text", "cc_emails",
  "fwd_emails", "reply_cc_emails"), include_requester = FALSE,
  include_stats = FALSE, include_custom_fields = FALSE,
  priorities_lookup = ticket_priorities, sources_lookup = ticket_sources,
  status_lookup = ticket_status, date_fields = ticket_date_fields)
}
\arguments{
\item{client}{The Freshdesk API client object (see \code{\link{freshdesk_client}}).}

\item{tickets_path}{The path of the tickets API. Defaults to \code{/api/v2/tickets}.
You should not need to change the default value.}

\item{remove_fields}{Fields returned by the Freshdesk API to remove from the data frame.
Defaults to removing the \code{description} and \code{description_text} fields
which could be quite verbose.}

\item{include_requester}{If \code{TRUE} returns additional attributes of the requester.}

\item{include_stats}{If \code{TRUE} returns additional attributes of the status.}

\item{include_custom_fields}{If \code{TRUE} customer fields will be included in the results.}

\item{priorities_lookup}{Optional dataframe of ticket priorities and associated values.
Defaults to \code{\link{ticket_priorities}} which is defined in the package.}

\item{sources_lookup}{Optional dataframe of ticket sources and associated values.
Defaults to \code{\link{ticket_sources}} which is defined in the package.}

\item{status_lookup}{Optional dataframe of ticket status and associated values.
Defaults to \code{\link{ticket_status}} which is defined in the package.}

\item{date_fields}{Fields returned by the Freshdesk API that are date fields. Defaults
to \code{\link{ticket_date_fields}} which is defined in the package.}
}
\value{
A data frame of tickets.
}
\description{
\code{tickets} queries the Freshdesk API and returns a data frame of ticke data.
}
\details{
This function queries the Freshdesk API to view the details regarding multiple
tickets.
}
\examples{
\dontrun{
fc <- freshdesk_client("your-domain", "your-api-key")

# view tickets
t <- tickets(fc)
}
}
