% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickets.R
\name{ticket}
\alias{ticket}
\title{View a Ticket}
\usage{
ticket(client, ticket_id, tickets_path = "/api/v2/tickets", include = NULL,
  valid_include_parameters = ticket_include_parameters,
  priorities_lookup = ticket_priorities, sources_lookup = ticket_sources,
  status_lookup = ticket_status)
}
\arguments{
\item{client}{The Freshdesk API client object (see \code{\link{freshdesk_client}}).}

\item{ticket_id}{The ID of the ticket you would like to view.}

\item{tickets_path}{The path of the tickets API. Defaults to \code{/api/v2/tickets}.
You should not need to change the default value.}

\item{include}{Optional include a vector of query string parameters.}

\item{valid_include_parameters}{Optional vector of valid include parameters. Defaults to
\code{\link{ticket_include_parameters}} which is defined in the package.}

\item{priorities_lookup}{Optional dataframe of ticket priorities and associated values.
Defaults to \code{\link{ticket_priorities}} which is defined in the package.}

\item{sources_lookup}{Optional dataframe of ticket sources and associated values.
Defaults to \code{\link{ticket_sources}} which is defined in the package.}

\item{status_lookup}{Optional dataframe of ticket status and associated values.
Defaults to \code{\link{ticket_status}} which is defined in the package.}
}
\value{
A list with the attributes of the ticket.
}
\description{
\code{ticket} queries the Freshdesk API and returns data related to a ticket.
}
\details{
This function queries the Freshdesk API based on ticket id to view the details
regarding this ticket. You can optionally include conversation, requester, company
or extra stats data.
}
\examples{
\dontrun{
fc <- freshdesk_client("your-domain", "your-api-key")

# view ticket with id = 3 and include addtional company and requester data
t <- ticket(fc, 3, include = "requester,company")
}
}
